Changed fileNames and titles
* Changed filename to match figma
* Changed title string to match figma
* Changed package name to match figma
