+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/somatichealth/SomaticHealthData.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.screens.triage.healthHistory
+package com.EENX15_22_17.digital_journal.android.screens.triage.somatichealth
 
 
 enum class SomaticHealthData {
